+++ CHANGELOG.md
@@ -6,6 +6,9 @@ ChangeLog history
 ## Deprecations 
 
 ## Features
+- add operator read-npy
+- add support to visualize points in neuroglancer
+- make read-json work with bbox in task
 
 ## Bug Fixes 
 - fix thresholding threshold type